+++ frontend/src/components/Navbar.jsx
@@ -2,16 +2,12 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import '../styles/navbar.css';
 
+
 export const Navbar = () => {
   return (
-    <nav className="navbar">
-      <div className="navbar-container">
-        <Link to="/" className="logo">Housing App</Link>
-        <ul className="nav-links">
-          <li><Link to="/">Listings</Link></li>
-          <li><Link to="/add-listing">Add Listing</Link></li>
-        </ul>
-      </div>
+    <nav style={{ background: '#007bff', color: '#fff', padding: '10px' }}>
+      <a href="/" style={{ marginRight: '10px', color: '#fff', textDecoration: 'none' }}>Home</a>
+      <a href="/add-listing" style={{ color: '#fff', textDecoration: 'none' }}>Add Listing</a>
     </nav>
   );
 };
