+++ frontend/src/pages/LoginPage.js
@@ -1,0 +1,4 @@
+const LoginPage = () => {
+  return <div>Login Page</div>;
+};
+export default LoginPage;
