--- frontend/src/pages/LoginPage.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-function LoginPage() {
-  return (
-    <div>
-      <h1>Login Page</h1>
-      <p>Login form will go here.</p>
-    </div>
-  );
-}
-
-export default LoginPage;
