+++ frontend/src/pages/RegisterPage.js
@@ -1,0 +1,4 @@
+const RegisterPage = () => {
+  return <div>Register Page</div>;
+};
+export default RegisterPage;
