--- frontend/src/pages/RegisterPage.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-function RegisterPage() {
-  return (
-    <div>
-      <h1>Register Page</h1>
-      <p>Registration form will go here.</p>
-    </div>
-  );
-}
-
-export default RegisterPage;
